pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "sim_server"
dynamic = ["version"]
description = "Jobrunner of Simrun Framework"
authors = [{ name = "Prajeesh Ag", email = "prajeeshag@gmail.com" }]
license = { file = "LICENSE" }
requires-python = ">=3.14,<3.15"
dependencies = [
    "pydantic>=2.12.5",
    "tortoise-orm[asyncpg,psycopg]>=1.1.7",
]

[tool.uv]
package = true

[project.scripts]
#sim_runner = "simrun.jobrunner.cli:app"

[project.optional-dependencies]


[tool.pdm.build]
excludes = ["tests", "examples"]

[tool.pdm.version]
source = "scm"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["sim_server"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.ruff]
lint.extend-select = ["T201"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-xdist>=3.8.0",
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "respx>=0.22.0",
]
